+++ APOYOS_SOCIALES/APOYOS_SOCIALES/Entities/Personal.cs
@@ -6,7 +6,7 @@
         public string Nombre { get; set; }
         public string Edad { get; set; } 
         public string FechaContratacion { get; set; } 
-        public int Salario { get; set; }
+        public decimal Salario { get; set; }
         public Genero Genero { get; set; }
         public Cargo Cargo { get; set; }
         public Area? Area { get; set; }
